Complete rewrite of the details scanner function
If a post which needed a details comment had one when the check was done
but had been alread removed in the same pass for breaking another rule,
then the comment giving reasoning for that posts removal would be
removed by the details_scan function. This has now been fixed by making
the function more efficient and a change in the order things happen.